files/requirements.txt
pandas==2.1.4
streamlit==1.31.0
plotly==5.18.0
sqlalchemy==2.0.25
mysql-connector-python==8.3.0
openpyxl==3.1.2
xlrd==2.0.1
apache-airflow==2.8.1
pytest==7.4.4
pytest-cov==4.1.0
flake8==7.0.0
